// app/src/store/modules/Listings/action-types.ts
export enum ListingsActionTypes {
  LISTINGS_REQUEST = "[Listings] LISTINGS_REQUEST",
  LISTINGS_SUCCESS = "[Listings] LISTINGS_SUCCESS",
  LISTINGS_FAILURE = "[Listings] LISTINGS_FAILURE",

  LISTING_CREATE_REQUEST = "[Listings] LISTING_CREATE_REQUEST",
  LISTING_CREATE_SUCCESS = "[Listings] LISTING_CREATE_SUCCESS",
  LISTING_CREATE_FAILURE = "[Listings] LISTING_CREATE_FAILURE",

  LISTING_DELETE_REQUEST = "[Listings] LISTING_DELETE_REQUEST",
  LISTING_DELETE_SUCCESS = "[Listings] LISTING_DELETE_SUCCESS",
  LISTING_DELETE_FAILURE = "[Listings] LISTING_DELETE_FAILURE"
}